Move the readinessCheck to framework/util
